+++ ResponsibleGherkin/ScenarioGenerator.cs
@@ -1,0 +1,26 @@
+using Gherkin.Ast;
+
+namespace ResponsibleGherkin;
+
+public static class ScenarioGenerator
+{
+	public static IEnumerable<Line> Generate(
+		Scenario scenario,
+		Flavor flavor,
+		GenerationContext context)
+	{
+		yield return $"[{flavor.TestAttribute}]";
+
+		var methodName = $"{scenario.Keyword.Trim()}_{scenario.Name.ToPascalCase()}";
+		yield return $"public {flavor.ReturnType} {methodName}() => this.{context.ExecutorName}.{flavor.RunMethod}(";
+
+		yield return GenerateScenario(scenario).IndentBy(1);
+
+		foreach (var line in StepsGenerator.Generate(scenario))
+		{
+			yield return line.IndentBy(1);
+		}
+	}
+
+	private static Line GenerateScenario(Scenario scenario) => $"Scenario({scenario.Name.Quote()}),";
+}
